allow empty header in RemoteBlockOutputStream #3411
The bug with inserts to Distributed tables was introduced in https://github.com/yandex/ClickHouse/pull/3171
It added a workaround specifically for inserting in the Native format without specifying the list of columns.
Native (as opposed to other formats) historically supports this. To signal that the input block structure
shouldn't conform to any fixed header in this case, the remote server started sending empty header block.
This commit adds support for empty headers to RemoteBlockOutputStream.