Support xxHash with external lz4
xxhash.h does not exist in external lz4 because it is an implementation detail
and its symbols are namespaced with XXH_NAMESPACE macro.  When lz4 is external,
xxHash has to be external too, and the dependents have to link to it.

This find_xxhash.cmake is based on find_lz4.cmake.