support reading from HA mode HDSF
the driver libhdfs3 can support HA mode. When the uri doesn't contain
port num, the libhdfs3 will handle it in HA mode, treat the uri as
nameservice name and read the real host and port from configuration file
as the origin Java client.
the default configuration file is hdfs-client.xml in the working
directory, it also can be set in env variable "LIBHDFS3_CONF".
the format of the configuration file is same with hdfs-site.xml.