Separate the hash table interface for ReverseIndex.
It is significantly different from other uses of hash tables, and poses
the main obstacle to changing the hash table interface to the one that
can be easily supported by compound hash tables.

Make it explicitly separate, implement it only for a particular kind of
HashTable, and move this implementation to the ReverseIndex module.