Fixed converting DateTime64-Field on client into DataTypeDateTime64-field type.
Since DateTime64 is just a typedef, and there is no explicit Field-type
for it, we have to solely rely on type_hint provided by origin column.
If the hint is missing, there is no way of distinguishing DateTime64
from Decimal64.

Alternative could be having explicit converting code, (and it looks like
it has to be added at some point).