Bind C-p/C-n to history-previous/history-next like readline in clickhouse-client with replxx
It is not easy to keep in mind that C-p/C-n is not
history-previous/history-next in replxx, thus it is pretty easy to
mistype.

Previous COMPLETE_NEXT/COMPLETE_PREV bindings has been binded to the
M-P/M-N (that was used for HISTORY_COMMON_PREFIX_SEARCH before, but it
also binded to M-p/M-n).

Plus clickhouse-client can be compiled with readline, so it is better to
make bindings the same.